plein de choses

+++ app/src/main/java/supdevinci/vitemeteo/model/HourlyWeather.kt
@@ -2,6 +2,6 @@ package supdevinci.vitemeteo.model
 
 data class HourlyWeather (
     val weatherCode: Int,
-    val temperature: Float,
+    val temperature: Int,
     val hour: String
     )
